+++ src/pages/Code.tsx
@@ -1,0 +1,11 @@
+const Code = () => {
+  return (
+    <>
+      <div className="min-h-full text-center text-5xl text-black dark:text-slate-300">
+        Code
+      </div>
+    </>
+  );
+};
+
+export default Code;
