+++ src/pages/Music.tsx
@@ -1,0 +1,9 @@
+const Music = () => {
+  return (
+    <div className="min-h-full text-center text-5xl text-black dark:text-slate-300">
+      Music
+    </div>
+  );
+};
+
+export default Music;
